games_service/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
python-jose==3.3.0
python-chess==1.999
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.20.0

